tests/fixtures/vulnerable-python-app/requirements.txt
Flask==2.3.0
Pillow==9.0.0
PyYAML==5.4
SQLAlchemy==2.0.0
Werkzeug==2.3.0
